\name{extractoR.utils-package}
\alias{extractoR.utils-package}
\alias{extractoR.utils}
\docType{package}
\encoding{UTF-8}
\title{CRAN packages metadata extractor (utils)}
\description{
  extractoR.utils is part of extractoR which is a set of R packages
  that can be used to fetch, extract and dump CRAN packages metadata in
  a SQL database. extractoR.utils contains utilities functions.
}

\author{
Maëlick Claes <himself@maelick.net>, Philippe Grosjean
<Philippe.Grosjean@umons.ac.be> & Prateek Rastogi
<p.rastogi@outlook.com>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\keyword{utils}
