\name{maintaineR-package}
\alias{maintaineR-package}
\alias{maintaineR}
\docType{package}
\encoding{UTF-8}
\title{maintaineR}
\description{
  maintaineR contains functions to analyse CRAN packages depencies,
  conflicts and clones.
}

\author{
Maëlick Claes <himself@maelick.net>

Maintainer: Maëlick Claes <himself@maelick.net>
}
\keyword{utils}
